requirements/dev.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-phonenumber-field==7.3.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
phonenumbers==8.13.35
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zope.interface==6.3
